# src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: istu_dwa_lb_2 API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
paths:
  /auth/login:
    post:
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [login, password]
              properties:
                login:
                  type: string
                password:
                  type: string
      responses:
        '200':
          description: OK
        '401':
          description: Invalid credentials
  /auth/me:
    get:
      summary: Current user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
  /auth/logout:
    post:
      summary: Logout
      responses:
        '200':
          description: OK
  /sections:
    get:
      summary: Menu sections
      responses:
        '200':
          description: OK
  /dishes:
    get:
      summary: List dishes
      parameters:
        - in: query
          name: sectionId
          schema:
            type: integer
        - in: query
          name: minWeight
          schema:
            type: integer
        - in: query
          name: maxWeight
          schema:
            type: integer
        - in: query
          name: minCalories
          schema:
            type: integer
        - in: query
          name: maxCalories
          schema:
            type: integer
        - in: query
          name: hasAllergens
          schema:
            type: string
            enum: ['true', 'false']
        - in: query
          name: isSpicy
          schema:
            type: string
            enum: ['true', 'false']
        - in: query
          name: kidFriendly
          schema:
            type: string
            enum: ['true', 'false']
        - in: query
          name: q
          schema:
            type: string
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
  /dishes/{id}:
    get:
      summary: Dish detail
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: OK
        '404':
          description: Not found
  /dishes/{id}/reviews:
    post:
      summary: Create dish review
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [authorName, body]
              properties:
                authorName:
                  type: string
                  maxLength: 255
                body:
                  type: string
                  maxLength: 8000
                photoUrl:
                  type: string
                  nullable: true
                  maxLength: 2048
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: integer
                  dishId:
                    type: integer
                  authorName:
                    type: string
                  body:
                    type: string
                  photoUrl:
                    type: string
                    nullable: true
                  createdAt:
                    type: string
                    format: date-time
        '400':
          description: Validation error
        '404':
          description: Dish not found
  /halls:
    get:
      summary: List halls
      responses:
        '200':
          description: OK
  /halls/{hallId}/tables:
    get:
      summary: Tables with availability
      parameters:
        - in: path
          name: hallId
          required: true
          schema:
            type: integer
        - in: query
          name: visitDate
          required: true
          schema:
            type: string
            format: date
      responses:
        '200':
          description: OK
        '400':
          description: Validation error
        '404':
          description: Hall not found
  /bookings:
    post:
      summary: Create booking
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tableId, visitDate, guestCount, contactName, contactPhone]
              properties:
                tableId:
                  type: integer
                visitDate:
                  type: string
                  format: date
                guestCount:
                  type: integer
                contactName:
                  type: string
                contactPhone:
                  type: string
                contactEmail:
                  type: string
                  nullable: true
      responses:
        '201':
          description: Created
        '400':
          description: Validation or capacity
        '404':
          description: Table not found
        '409':
          description: Conflict
  /admin/bookings:
    get:
      summary: Admin list bookings
      description: Только заголовок Authorization Bearer (cookie не учитывается).
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /admin/bookings/{id}:
    delete:
      summary: Admin delete booking
      description: Только заголовок Authorization Bearer (cookie не учитывается).
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
  /admin/reviews:
    get:
      summary: Admin list reviews
      description: Только заголовок Authorization Bearer (cookie не учитывается).
      security:
        - bearerAuth: []
      parameters:
        - in: query
          name: limit
          schema:
            type: integer
            minimum: 1
            maximum: 200
      responses:
        '200':
          description: OK
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
  /admin/reviews/{id}:
    delete:
      summary: Admin delete review
      description: Только заголовок Authorization Bearer (cookie не учитывается).
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        '204':
          description: Deleted
        '401':
          description: Unauthorized
        '403':
          description: Forbidden
        '404':
          description: Not found
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT